Add command to list supported currencies
By using "mmc list" the user can now
get a list of all currencies supported
my mmc. This helps in debugging
usage errors and offers the user to get
an overview of available currency IDs
as they might not always be obvious.
It also provides a way to give inside to the
applications ample capabilities.

Signed-off-by: David <david@isan.engineer>